requirements.txt
snowflake-connector-python[pandas]
pandas
sqlalchemy
psycopg2-binary
dbt-snowflake
apache-airflow 